:recycle: refactor: Answer staff인 사람만 할 수 있게 수정 (#27)

+++ src/main/java/com/smunity/graduation/domain/qna/service/AnswerService.java
@@ -1,14 +1,11 @@
 package com.smunity.graduation.domain.qna.service;
 
 import com.smunity.graduation.domain.accounts.entity.User;
-import com.smunity.graduation.domain.accounts.exception.AccountsExceptionHandler;
-import com.smunity.graduation.domain.accounts.repository.user.UserRepository;
 import com.smunity.graduation.domain.qna.dto.AnswerRequestDto;
 import com.smunity.graduation.domain.qna.dto.AnswerResponseDto;
 import com.smunity.graduation.domain.qna.entity.Answer;
 import com.smunity.graduation.domain.qna.entity.Question;
 import com.smunity.graduation.domain.qna.repository.AnswerJpaRepository;
-import com.smunity.graduation.global.common.code.status.ErrorCode;
 import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -19,20 +16,19 @@ import org.springframework.stereotype.Service;
 public class AnswerService {
 
     private final AnswerJpaRepository answerJpaRepository;
-    private final UserRepository userRepository;
     private final QnAServiceUtils qnAServiceUtils;
 
     public AnswerResponseDto createAnswer(Long questionId, AnswerRequestDto requestDto) {
-        User author = userRepository.findByUserName("admin")
-                .orElseThrow(() -> new AccountsExceptionHandler(ErrorCode.USER_NOT_FOUND));
-
+        User author = qnAServiceUtils.getCurrentUser();
         Question question = qnAServiceUtils.getQuestionById(questionId);
+        qnAServiceUtils.validateStaffAccess();
         Answer saveAnswer = answerJpaRepository.save(new Answer(null, requestDto.content(), author, question));
         return AnswerResponseDto.from(saveAnswer);
     }
 
     public AnswerResponseDto updateAnswer(Long answerId, AnswerRequestDto requestDto) {
         Answer existingAnswer = qnAServiceUtils.getAnswerById(answerId);
+        qnAServiceUtils.validateStaffAccess();
         existingAnswer.setContent(requestDto.content());
         Answer updateAnswer = answerJpaRepository.save(existingAnswer);
         return AnswerResponseDto.from(updateAnswer);
@@ -40,6 +36,7 @@ public class AnswerService {
 
     public void deleteAnswer(Long answerId) {
         Answer existingAnswer = qnAServiceUtils.getAnswerById(answerId);
+        qnAServiceUtils.validateStaffAccess();
         answerJpaRepository.delete(existingAnswer);
     }
 }
